Create Closest(Next)PrimeorFib.py
Given a number n, find the number starting from n which is either prime or Fibonacci. The first such number that is either prime or Fibonacci is the answer.

Hint: Write two functions isFibo and isPrime. Starting from n keep on checking till you find a number that is either Fib or prime.

Prime numbers are 2 3 5 7 11 etc.

Fib numbers are 0 1 1 2 3 5 8 13 etc.