Create farthestCoprime.py
Given an array of n numbers, replace each element with it’s farthest coprime in the range [2, 250]. Example, the farthest coprime for 2 is 249 and for 243 is 2. Two numbers are said to be coprimes if they have no common factor except for 1. All elements are within the range 2 to 250, both included and for any number there will be only one farthest coprime.