STRING.py
Write a program which accepts a string and returns a string made of the first 2 and the last 2 characters of the given string. If the string length is less than 2, return -1.

so, for input w3resource, you should print w3ce and for input B, you should print -1 as the length of input is less than 2

Note: If the string length is equal to 2, consider the 2 characters to be the first as well as the last two characters.